source/back-end/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
chromadb==1.0.9
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
Deprecated==1.2.18
distro==1.9.0
durationpy==0.9
fastapi==0.115.9
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.3.2
ftfy==6.3.1
google-auth==2.40.1
googleapis-common-protos==1.70.0
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.31.2
humanfriendly==10.0
ibm-cos-sdk==2.14.1
ibm-cos-sdk-core==2.14.1
ibm-cos-sdk-s3transfer==2.14.1
ibm_watsonx_ai==1.3.20
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kubernetes==32.0.1
lomond==0.3.3
markdown-it-py==3.0.0
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
numpy==2.2.5
oauthlib==3.2.2
onnxruntime==1.22.0
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
orjson==3.10.18
overrides==7.7.0
packaging==25.0
pandas==2.2.3
posthog==4.0.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.2
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.25.0
rsa==4.9.1
safetensors==0.5.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.51.3
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
zipp==3.21.0
